LayoutTests/editing/pasteboard/paste-text-018.html

<html> 
<head>
<style>
.editing { 
    border: 2px solid red; 
    font-size: 24px; 
}
.explanation { 
    border: 2px solid blue; 
    padding: 12px; 
    font-size: 24px; 
    margin-bottom: 24px;
}
.scenario { margin-bottom: 16px;}
.scenario:first-line { font-weight: bold; margin-bottom: 16px;}
.expected-results:first-line { font-weight: bold }
</style>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
<title>Editing Test</title> 
</head> 
<body>
<div class="explanation">
<div class="scenario">
Tests: 
<br>
Fix for this bug: 
<a href="rdar://problem/4039676">&lt;rdar://problem/4039676&gt;</a> REGRESSION (Mail): in reply, <cr> on pasted text (without trailing newline) is eaten when pasted above date
</div>
<div class="expected-results">
Expected Results:
<br>
Should see two lines of text below, self-documenting themselves as "line one" and "line two".
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="test" class="editing">
line one<br>line two
</div>
</div>
<script>

Markup.description(document.getElementsByClassName('explanation')[0].textContent);

function editingTest() {
    for (i = 0; i < 8; i++)
        extendSelectionForwardByCharacterCommand();
    cutCommand();
    Markup.dump('root', 'After copy');
    pasteCommand();
    Markup.dump('root', 'After paste');
}

runEditingTest();
</script>
</body>
</html>
